pytorch_ddp/geo/requirements.txt
torchgeo
transformers
torchvision
pillow
soxr
